refactor: update bin name, dependencies; remove benches except all
