+++ Makefile.toml
@@ -95,23 +95,11 @@ command = "cargo"
 args = ["build", "--no-default-features", "--target", "wasm32-unknown-unknown", "--workspace", "--exclude", "miden-mock"]
 
 # --- benchmarking --------------------------------------------------------------------------------
-[tasks.bench-all]
+[tasks.bench-tx]
 description = "Run all available transaction benchmarks"
 workspace = false
 command = "cargo"
-args = ["run", "--bin", "benchmarks", "--features", "executable", "all"]
-
-[tasks.bench-tx]
-description = "Run the transaction benchmark with default transaction script and notes"
-workspace = false
-command = "cargo"
-args = ["run", "--bin", "benchmarks", "--features", "executable", "simple"]
-
-[tasks.bench-p2id]
-description = "Run the transaction benchmark which consumes a P2ID note into a basic wallet"
-workspace = false
-command = "cargo"
-args = ["run", "--bin", "benchmarks", "--features", "executable", "p2id"]
+args = ["run", "--bin", "bench-tx"]
 
 # --- utilities -----------------------------------------------------------------------------------
 [tasks.watch]
