+++ bench-tx/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "miden-bench-tx"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+[[bin]]
+name = "bench-tx"
+path = "src/main.rs"
+
+[dependencies]
+miden-lib = { package = "miden-lib", path = "../miden-lib", version = "0.3" }
+miden-objects = { package = "miden-objects", path = "../objects", version = "0.3" }
+miden-tx = { package = "miden-tx", path = "../miden-tx", version = "0.3" }
+mock = { package = "miden-mock", path = "../mock", features = ["std"] }
+serde_json = { package = "serde_json", version = "1.0" }
+vm-processor = { workspace = true }
