--- benchmarks/src/results/bench_p2id.json
@@ -1,13 +1,0 @@
-{
-  "p2id": {
-    "epilogue": 270,
-    "note_execution": [
-      {
-        "0xb9fa30eb43d80d579be02dc004338e06b5ad565e81e0bac11a94ab01abfdd40a": 881
-      }
-    ],
-    "notes_processing": 918,
-    "prologue": 2002,
-    "tx_script_processing": 88207
-  }
-}
