+++ miden-tx/Cargo.toml
@@ -29,5 +29,5 @@ miden-verifier = { workspace = true }
 vm-processor = { workspace = true }
 
 [dev-dependencies]
-rand_chacha = { version = "0.3", default-features = false }
 mock = { package = "miden-mock", path = "../mock", default-features = false }
+rand_chacha = { version = "0.3", default-features = false }
